requirements.txt
# GUI フレームワーク
flet>=0.21.2

# ドキュメント処理
python-docx>=1.1.0
PyPDF2>=3.0.1
mammoth>=1.6.0  # docxのリッチテキスト抽出用（オプション）

# AI API クライアント
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# HTTP クライアント
httpx>=0.27.0
aiohttp>=3.9.0

# ユーティリティ
tiktoken>=0.6.0  # トークンカウント用
python-dotenv>=1.0.0  # 環境変数読み込み用

# その他
asyncio-throttle>=1.0.2  # API レート制限対応
pydantic>=2.5.0  # データ検証用